Try with sanboot
(didn't work)

+++ files/boot.py
@@ -14,6 +14,8 @@ syslog.setlogmask(syslog.LOG_UPTO(syslog.LOG_INFO))
 # from the name, e.g. c1-3.cloud.example.org take c1-3
 hostname = socket.gethostbyaddr(os.environ["REMOTE_ADDR"])[0].split(".")[0]
 syslog.syslog(syslog.LOG_DEBUG, "Got boot iPXE request from " + hostname)
+
+started = False
 try:
     os.stat("/var/www/provision/memtest86/" + hostname)
     os.remove("/var/www/provision/memtest86/" + hostname)
@@ -31,48 +33,48 @@ try:
 
     f.close()
     print "#!ipxe"
-    print "sanboot nodesettings['memtest86_iso_path']"
+    print "sanboot " + nodesettings['memtest86_iso_path']
+    started = True
 
 # Catch the exception when the memtest file wasn't found
-# TODO: retry pxe in case it has reinstall?
 except OSError:
-    print "#!ipxe"
-    print "exit"
+    pass
 # Catch all other problems
 except Exception as exc:
-    print str(exc)
+    #print str(exc)
     syslog.syslog(syslog.LOG_ERR, str(exc))
 
 ########
 
-try:
-    os.stat("/var/www/provision/reinstall/" + hostname)
-    os.remove("/var/www/provision/reinstall/" + hostname)
-    f = open("/var/www/provision/nodes/" + hostname + ".conf")
+if started == False:
+    try:
+        os.stat("/var/www/provision/reinstall/" + hostname)
+        os.remove("/var/www/provision/reinstall/" + hostname)
+        f = open("/var/www/provision/nodes/" + hostname + ".conf")
 
-    syslog.syslog(syslog.LOG_INFO, "Reinstalling node " + hostname)
-    nodesettings = {}
-    for line in f.readlines():
-      #for every line, e.g. "key=value", set nodesettings["key"]="value"
-      #comment lines will throw an error, skip them
-        try:
-            nodesettings[line.split("=")[0]] = line.split("=", 1)[1].strip()
-        except:
-            pass
+        syslog.syslog(syslog.LOG_INFO, "Reinstalling node " + hostname)
+        nodesettings = {}
+        for line in f.readlines():
+          #for every line, e.g. "key=value", set nodesettings["key"]="value"
+          #comment lines will throw an error, skip them
+            try:
+                nodesettings[line.split("=")[0]] = line.split("=", 1)[1].strip()
+            except:
+                pass
 
-    f.close()
-    print "#!ipxe"
-    print "kernel " + nodesettings["kernel_url_path"] + "/vmlinuz ks=" + nodesettings["kickstart_url"] + " edd=off ksdevice=bootif kssendmac console=ttyS1,115200 console=tty0 initrd=initrd.img " + nodesettings.get("extra_kernel_params", "")
-    print "initrd " + nodesettings["kernel_url_path"] + "/initrd.img"
-    print "boot"
+        f.close()
+        print "#!ipxe"
+        print "kernel " + nodesettings["kernel_url_path"] + "/vmlinuz ks=" + nodesettings["kickstart_url"] + " edd=off ksdevice=bootif kssendmac console=ttyS1,115200 console=tty0 initrd=initrd.img " + nodesettings.get("extra_kernel_params", "")
+        print "initrd " + nodesettings["kernel_url_path"] + "/initrd.img"
+        print "boot"
 
-# Catch the exception when the reinstall file wasn't found
-except OSError:
-    print "#!ipxe"
-    print "exit"
-# Catch all other problems
-except Exception as exc:
-    print str(exc)
-    syslog.syslog(syslog.LOG_ERR, str(exc))
+    # Catch the exception when the any memtest/reinstall file wasn't found
+    except OSError:
+        print "#!ipxe"
+        print "exit"
+    # Catch all other problems
+    except Exception as exc:
+        print str(exc)
+        syslog.syslog(syslog.LOG_ERR, str(exc))
 
 syslog.closelog()
